silence compiler warnings

+++ src/EXTRA-DUMP/dump_xtc.cpp
@@ -381,7 +381,7 @@ static int *buf = nullptr;
  | with some routines to assist in this task (those are marked
  | static and cannot be called from user programs)
 */
-#define MAXABS (INT_MAX-2)
+#define MAXABS (float)(INT_MAX-2)
 
 #ifndef SQR
 #define SQR(x) ((x)*(x))
